Add securest CSP headers for lando app
